// libs/content-ee/governance/src/lib/record/models/declare-record.model.ts
/*
 * Copyright © 2005-2024 Hyland Software, Inc. and its affiliates. All rights reserved.
 *
 * License rights for this program may be obtained from Hyland Software, Inc.
 * pursuant to a written agreement and any use of this program without such an
 * agreement is prohibited.
 */

import { Node } from '@alfresco/js-api';
export interface DeclareRecord {
    entry: Node;
    status: 'failed' | 'passed';
}